Move some `is*` meths to `tools\variable\analyzer`.
The following methods have been moved to `tools\variable\analyzer`:
  * `test::isRegex`,
  * `test::isValidIdentifier`,
  * `test::isValidNamespace`.
